Refactor the Java build to use a more conventional module style
Create a top level Maven parent and nest the modules under that.  Rename the modules
to match the Quiver project and allow the various JMS clients to bring in their own
version of the JMS API making the quiver-jms-driver module only use one for compile
and depending on the client module to bring in the one the actual client needs.