re-use message in qpid-proton-cpp and qpid-messaging-cpp arrows
Allocating a message for each send is a big overhead, it is more efficient to
create one and just update the timestamp/id each time. The old qpid-cpp test
tools all did this.

Before fixing this, the proton-cpp arrow was a little slower than qpid-messaging-cpp.
Afterwards, both are faster but proton-cpp is significantly faster than qpid-messaging!