Fix interop problems, add interop test.
go-arrow was encoding the SendTime property as an AMQP timestamp, other arrows
encode as AMQP LONG in milliseconds since the epoch.

Arrows were inconsistent about encoding of message-id, some used long, others string.
Made c, go, python and cpp consistently use a string representation of a counter
ID.

scripts/test-interop script to verify all combinations of c, go, python and cpp
with a broker/router. Ideally would use peer-to-peer to avoid the need for a
broker, but c++ arrow is not working as a passive server receiver.