Provide a Netty tc_native version specific to the target library
Due to issues of incompatibility between ancient netty versions that
the Artemis client and the slight lag in the other projects catching
up to each other I have opted to rely on the project specifc netty
version include a matching netty-tcnative version when updating the
client version.  This seems to result in less breakage from Artemis
and tests each client under the version assumptions relative to the
release being used.